lexcredendi/requirements.txt
beautifulsoup4==4.12.3
Django==5.1.3
django-cors-headers==4.3.1
django-import-export==3.3.8
django-tz-detect==5.1.0
gunicorn==22.0.0
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0
requests==2.31.0
mysqlclient==2.2.4